pyproject.toml
[tool.poetry]
name = "subgroup-separability"
version = "1.0.0"
description = ""
readme = "README.md"
packages = [{include = "subgroup_separability"}]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^1.13.1"
torchmetrics = "^0.11.0"
torchvision = "^0.14.1"
tqdm = "^4.64.1"
hydra-core = "^1.3.1"
wandb = "^0.13.9"
pandas = "^1.5.3"
ipykernel = "^6.20.2"
scikit-learn = "^1.2.0"
matplotlib = "^3.6.3"
schema = "^0.7.5"
fastapi = "^0.86.0"
pytorch-lightning = "^1.9.0"
tensorboardx = "^2.5.1"
openpyxl = "^3.1.0"
seaborn = "^0.12.2"
scikit-image = "^0.19.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
gpustat = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
